Create MontyHall
The monty hall problem centers around a gameshow like game where the contestant chooses one of three doors. Behind one door is a prize. The other two doors do not have a prize. After selecting a door the host opens a door the contestant did not choose that is a losing door. The contestant then has the opportunity to change their selected door. Should they change it? Most people assume that since there are 2 doors and one door has a prize it doesn't matter if they switch as their odds of winning is 50%. This program simulates this decision if the player always chooses to switch doors after the reveal of one incorrect guess. The number of losses displays is the number of wins if the contestant chooses not to change their guess. The program runs the decision 100,000,000 times.